+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHKE3T0ZR5BKFM256MXKE1RW"
+        "01KHKJDHR3XYKN0SN71SMNNJMR"
     )
     r"""The unique identifier of the data source"""