+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHKE3T14HWA7C191G0SH0RJM"
+        "01KHKJDHR6KDS4KCV9CNDMJWKE"
     )
     r"""The unique identifier of the data source"""